// src/data/auth/types.ts
/* eslint-disable @typescript-eslint/no-explicit-any */
import { LogoutOptions, PopupLoginOptions, RedirectLoginOptions } from '@auth0/auth0-react'

export type ActionMapType<M extends { [index: string]: any }> = {
  [Key in keyof M]: M[Key] extends undefined
    ? {
        type: Key
      }
    : {
        type: Key
        payload: M[Key]
      }
}

export type AuthUserType = null | Record<string, any>

export type AuthStateType = {
  status?: string
  loading: boolean
  user: AuthUserType
}

type CanRemove = {
  login?: (email: string, password: string) => Promise<void>
  register?: (email: string, password: string, firstName: string, lastName: string) => Promise<void>
  //
  loginWithGoogle?: () => Promise<void>
  loginWithGithub?: () => Promise<void>
  loginWithTwitter?: () => Promise<void>
  //
  loginWithPopup?: (options?: PopupLoginOptions) => Promise<void>
  loginWithRedirect?: (options?: RedirectLoginOptions) => Promise<void>
  //
  confirmRegister?: (email: string, code: string) => Promise<void>
  forgotPassword?: (email: string) => Promise<void>
  resendCodeRegister?: (email: string) => Promise<void>
  newPassword?: (email: string, code: string, password: string) => Promise<void>
}

export type JWTContextType = CanRemove & {
  user: AuthUserType
  method: string
  loading: boolean
  authenticated: boolean
  unauthenticated: boolean
  login: (email: string, password: string) => Promise<void>
  register: (email: string, password: string, firstName: string, lastName: string) => Promise<void>
  logout: () => Promise<void>
}

export type FirebaseContextType = CanRemove & {
  user: AuthUserType
  method: string
  loading: boolean
  authenticated: boolean
  unauthenticated: boolean
  logout: () => Promise<void>
  loginWithGoogle: () => Promise<void>
  loginWithGithub: () => Promise<void>
  loginWithTwitter: () => Promise<void>
  forgotPassword?: (email: string) => Promise<void>
  login: (email: string, password: string) => Promise<void>
  register: (email: string, password: string, firstName: string, lastName: string) => Promise<void>
}

export type AmplifyContextType = CanRemove & {
  user: AuthUserType
  method: string
  loading: boolean
  authenticated: boolean
  unauthenticated: boolean
  login: (email: string, password: string) => Promise<unknown>
  register: (email: string, password: string, firstName: string, lastName: string) => Promise<unknown>
  logout: () => Promise<unknown>
  confirmRegister: (email: string, code: string) => Promise<void>
  forgotPassword: (email: string) => Promise<void>
  resendCodeRegister: (email: string) => Promise<void>
  newPassword: (email: string, code: string, password: string) => Promise<void>
}

export type Auth0ContextType = CanRemove & {
  user: AuthUserType
  method: string
  loading: boolean
  authenticated: boolean
  unauthenticated: boolean
  loginWithPopup: (options?: PopupLoginOptions) => Promise<void>
  loginWithRedirect: (options?: RedirectLoginOptions) => Promise<void>
  logout: (options?: LogoutOptions) => Promise<void>
}